Safari-609.1.11.1/LayoutTests/fast/scrolling/programmatic-scroll-to-negative-offset-expected.txt

Tests a programmatic scroll to a negative offset does not result in negative scrollLeft/scrollTop

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.scrollLeft is 0
PASS scroller.scrollLeft is 0
PASS successfullyParsed is true

TEST COMPLETE
